+++ hybrid-homework/web/src/commons/layout/footer/constants.ts
@@ -1,0 +1,10 @@
+// prettier-ignore
+// export const FOOTER_OPTIONS = (params) => ({
+//   "/solplace-logs": {isButton: false, isNav: true, isFixed: true},
+//   "/solplace-logs/new": { },
+//   "/solplace-logs/new/map": {},
+//   [`/solplace-logs/${params.solplaceLogId}`]: {},
+//   [`/solplace-logs/${params.solplaceLogId}/edit`]: {},
+//   [`/solplace-logs/${params.solplaceLogId}/edit/map`]: {},
+//   "/mypages": {isButton: false, isNav: true, isFixed: true}
+// })
