+++ hybrid-homework/web/src/components/commons/add-image/index.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import Image from "next/image";
 import styles from "./styles.module.css";
 import { ChangeEvent, useRef, useState } from "react";
