+++ hybrid-homework/web/src/components/commons/address-bottom/index.tsx
@@ -2,7 +2,7 @@
 
 import { useRouter, useSearchParams } from "next/navigation";
 import styles from "./styles.module.css";
-import { ButtonPrimaryMFull } from "@/commons/ui/button";
+import { ButtonPrimaryMFull } from "@/components/commons/button";
 
 interface AddressBottomProps {
   isEdit: boolean;
@@ -11,15 +11,15 @@ interface AddressBottomProps {
 export default function AddressBottom({ isEdit }: AddressBottomProps) {
   const router = useRouter();
   const searchParams = useSearchParams();
-  const redirectUrl = searchParams.get("redirectUrl");
+  const redirect = searchParams.get("redirect");
   const address = searchParams.get("address") || "";
-  const name = searchParams.get("name");
+  const title = searchParams.get("title");
   const contents = searchParams.get("contents");
 
   // 지도등록  지도수정 눌렀을 시 redirect-url 로 쿼리스트링과 함꼐 페이지 이동
   const onClickAddMap = () => {
     router.push(
-      `${redirectUrl}?name=${name}&contents=${contents}&address=${address}`
+      `${redirect}?title=${title}&contents=${contents}&address=${address}`
     );
   };
