+++ hybrid-homework/web/src/components/commons/form/index.tsx
@@ -1,0 +1,22 @@
+"use client";
+
+import styles from "./styles.module.css";
+import { FormProvider, useForm } from "react-hook-form";
+import { zodResolver } from "@hookform/resolvers/zod";
+
+export default function Form({ children, schema, useInitialize }) {
+  const methods = useForm({
+    resolver: zodResolver(schema),
+    mode: "onChange",
+  });
+
+  const { onSubmit } = useInitialize(methods);
+
+  return (
+    <FormProvider {...methods}>
+      <form className={styles.form} onSubmit={methods.handleSubmit(onSubmit)}>
+        {children}
+      </form>
+    </FormProvider>
+  );
+}
