+++ hybrid-homework/web/src/components/commons/form/styles.module.css
@@ -1,0 +1,6 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  gap: 2.5rem;
+}
