+++ hybrid-homework/web/src/components/solplace-logs-detail-edit/form.schema.ts
@@ -1,7 +1,9 @@
+"use client";
+
 import { z } from "zod";
 
 export const schema = z.object({
-  name: z.string().min(1),
+  title: z.string().min(1),
   contents: z.string().min(1),
   address: z.string(),
 });
